CDI-637 rename the builder package to configurator and fix some language.
