+++ api/src/main/java/javax/enterprise/inject/spi/AfterBeanDiscovery.java
@@ -17,8 +17,8 @@
 package javax.enterprise.inject.spi;
 
 import javax.enterprise.context.spi.Context;
-import javax.enterprise.inject.spi.builder.BeanConfigurator;
-import javax.enterprise.inject.spi.builder.ObserverMethodConfigurator;
+import javax.enterprise.inject.spi.configurator.BeanConfigurator;
+import javax.enterprise.inject.spi.configurator.ObserverMethodConfigurator;
 
 /**
  * <p>
