+++ api/src/main/java/javax/enterprise/inject/spi/AfterTypeDiscovery.java
@@ -17,7 +17,7 @@
 
 package javax.enterprise.inject.spi;
 
-import javax.enterprise.inject.spi.builder.AnnotatedTypeConfigurator;
+import javax.enterprise.inject.spi.configurator.AnnotatedTypeConfigurator;
 import java.util.List;
 
 /**
