+++ api/src/main/java/javax/enterprise/inject/spi/ProcessAnnotatedType.java
@@ -16,7 +16,7 @@
  */
 package javax.enterprise.inject.spi;
 
-import javax.enterprise.inject.spi.builder.AnnotatedTypeConfigurator;
+import javax.enterprise.inject.spi.configurator.AnnotatedTypeConfigurator;
 
 /**
  * <p>
