+++ api/src/main/java/javax/enterprise/inject/spi/ProcessObserverMethod.java
@@ -17,7 +17,7 @@
 
 package javax.enterprise.inject.spi;
 
-import javax.enterprise.inject.spi.builder.ObserverMethodConfigurator;
+import javax.enterprise.inject.spi.configurator.ObserverMethodConfigurator;
 
 /**
  * <p>
