+++ api/src/main/java/javax/enterprise/inject/spi/configurator/AnnotatedTypeConfigurator.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package javax.enterprise.inject.spi.builder;
+package javax.enterprise.inject.spi.configurator;
 
 import javax.enterprise.inject.spi.AnnotatedConstructor;
 import javax.enterprise.inject.spi.AnnotatedField;
