+++ api/src/main/java/javax/enterprise/inject/spi/configurator/InjectionPointConfigurator.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package javax.enterprise.inject.spi.builder;
+package javax.enterprise.inject.spi.configurator;
 
 import javax.enterprise.inject.spi.AnnotatedField;
 import javax.enterprise.inject.spi.AnnotatedParameter;
