+++ api/src/main/java/javax/enterprise/inject/spi/configurator/ObserverMethodConfigurator.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package javax.enterprise.inject.spi.builder;
+package javax.enterprise.inject.spi.configurator;
 
 import javax.enterprise.event.Reception;
 import javax.enterprise.event.TransactionPhase;
